Version: 2012-05-31 93u+
Sourced from ast-open-history version 2012-05-31 and:
7b97b44798d6fa664434b53f0f5e48b98103f25108c1d03e7d58dedf2ab7af0d  INIT.2012-05-31.tgz

+++ src/lib/libtksh/src/list.c
@@ -261,7 +261,7 @@ typedef struct TkshRegExp
 {
 	char *re;
 	char *str;
-	int pos[20];
+	ssize_t pos[20];
 	int n;
 } TkshRegExp;
